main.py
'''
Main execution entry point for the compliance QA pipeline
'''
import uuid
import json
import logging
from pprint import pprint

from dotenv import load_dotenv
load_dotenv(override=True)

from backend.src.graph.workflow import app

logging.basicConfig(
    level = logging.INFO,
    format="%(asctime)s - %(levelname)s - %(message)s" 
)
logger = logging.getLogger("youtube-add-compliance-checker-runner")


def run_cli_simulation():
    '''
    Simulates the video compliance audit request
    '''

    # generate the session ID
    session_id = str(uuid.uuid4())
    logger.info(f"Starting Audit Session : {session_id}")

    # define the initial state 
    initial_inputs = {
        "video_url" : "https://youtu.be/oGSFRxb0DM8",
        "video_id" : f"vid_{session_id[:8]}",
        "compliance_results" : [],
        "errors" : []
    }

    print("......initializing workflow.......")
    print(f"Input Payload : {json.dumps(initial_inputs, indent=2)}")

    try:
        final_state = app.invoke(initial_inputs)
        print("\n......workflow execution is complete......")

        print("\n Compliance Audit Report==")
        print(f"Video ID : {final_state.get('video_id')}")
        print(f"Status : {final_state.get('final_status')}")
        print("\n [VIOLATIONS DETECTED]")
        results = final_state.get('compliance_results',[])
        if results:
            for issue in results:
                print(f"- [{issue.get('severity')}] [{issue.get('category')}] : [{issue.get('description')}]")
        else:
            print("No violations detected......")
        print("\n[FINAL SUMMARY]")
        print(final_state.get('final_report'))

    except Exception as e:
        logger.error(f"Workflow Execution Failed : {str(e)}")
        raise e
    

if __name__ == "__main__":
    run_cli_simulation()

    
