requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
azure-common==1.1.28
azure-core==1.39.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.3
azure-monitor-opentelemetry==1.8.7
azure-monitor-opentelemetry-exporter==1.0.0b50
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.135.3
firecrawl-py==4.22.0
frozenlist==1.8.0
gitdb==4.0.12
gitpython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
isodate==0.7.2
jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
markupsafe==3.0.3
marshmallow==3.26.2
msal==1.35.1
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
mypy-extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-psycopg2==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-urllib==0.61b0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydeck==0.9.1
pyjwt==2.12.1
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sqlalchemy==2.0.49
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2026.1
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.43.0
watchdog==6.0.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2026.3.17
zipp==3.23.0
zstandard==0.25.0